Move test_valve functions to separate library.
